[REFACTOR] count 타입 변경 #S09P31C207-138

+++ server/ploggingservice/src/main/java/com/plonit/ploggingservice/api/plogging/controller/request/GrantMemberBadgeReq.java
@@ -8,7 +8,7 @@ import lombok.*;
 @Builder
 public class GrantMemberBadgeReq {
 
-    private Integer ploggingCount;
+    private Long ploggingCount;
 
     private Double distance;
